backend/requirements.txt
# Core API dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0

# Image processing
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78

# AI/ML dependencies (optional for basic functionality)
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
diffusers>=0.24.0
controlnet-aux>=0.4.0
accelerate>=0.24.0
xformers>=0.0.22

# Database (for future features)
sqlalchemy==2.0.23
alembic==1.12.1

# Security (for future features)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
